# perf: pre-allocate text change string capacity