Pretty trivial to maintain this now.

```rs
use deno_ast::apply_text_changes;
use deno_ast::TextChange;

fn main() {
  let mut value = 0;
  let mut large_string = String::new();
  for _ in 0..10000 {
    large_string.push_str("testing this out");
  }
  let changes = vec![TextChange::new(9, 10, "z".to_string()), TextChange::new(4, 6, "y".to_string()), TextChange::new(1, 2, "x".to_string())];
  for _ in 0..10000 {
    let final_str = apply_text_changes("testing this out", changes.clone());
    value += final_str.len();
  }
  eprintln!("start: {}", value);
}
```

```
> hyperfine --warmup 4 "before.exe" "after.exe"
Benchmark 1: before.exe
  Time (mean ± σ):       6.9 ms ±   0.6 ms    [User: 1.1 ms, System: 1.0 ms]
  Range (min … max):     5.9 ms …   9.5 ms    259 runs

Benchmark 2: after.exe
  Time (mean ± σ):       6.4 ms ±   0.5 ms    [User: 0.2 ms, System: 0.8 ms]
  Range (min … max):     5.6 ms …   8.5 ms    302 runs

Summary
  after.exe ran
    1.08 ± 0.12 times faster than before.exe
```

Closes https://github.com/denoland/deno_ast/issues/140